refactor(framework): break the server/telefunc-serve import cycle via a leaf types module
